Moved common slider components (#45)
