+++ src/routes/our-music/VolumeSlider.svelte
@@ -1,59 +1,16 @@
 <script>
-    import { onMount, tick } from 'svelte';
-    import { createEventDispatcher } from 'svelte';
-    const dispatch = createEventDispatcher();
-
     export let min;
     export let max;
     export let step;
     export let current;
+    let size = 13;
+    let sliderWidth = 85;
+    let fillColor = '#d6d9c7';
 
-    let sliding = false;
-    let slider;
-    let scale;
+    import { createEventDispatcher } from 'svelte';
+    const dispatch = createEventDispatcher();
 
-    onMount(() => {
-        scale = slider.clientWidth / (max / step);
-        function handleResize() {
-            scale = slider.clientWidth / (max / step);
-        }
-        window.addEventListener('resize', handleResize);
-        return () => window.removeEventListener('resize', handleResize);
-    });
-
-    $: dispatch('change', { value: current });
-
-    function handleMouseMove(e) {
-        if (sliding) {
-            const distance = e.clientX - slider.getBoundingClientRect().left;
-            const value = Math.round(distance / scale) * step;
-            current = Math.max(Math.min(value, max), min);
-        }
-    }
-
-    async function handleMouseDown(e) {
-        sliding = true;
-
-        const distance = e.clientX - slider.getBoundingClientRect().left;
-        const value = Math.round(distance / scale) * step;
-        current = value;
-    }
-
-    function handleKeyDown(e) {
-        if (e.keyCode === 37) {
-            e.preventDefault();
-            const nextValue = current - step;
-            current = Math.max(nextValue, min);
-        } else if (e.keyCode === 39) {
-            e.preventDefault();
-            const nextValue = current + step;
-            current = Math.min(nextValue, max);
-        }
-    }
-
-    function handleMouseUp(e) {
-        sliding = false;
-    }
+    import Slider from './Slider.svelte';
 
     export let muted;
     const logo = {
@@ -69,50 +26,10 @@
         justify-content: space-between;
         align-items: center;
     }
-
     .slidecontainer label {
         display: flex;
         text-align: center;
     }
-
-    .slider {
-        position: relative;
-        width: 85%;
-        height: 30px;
-        cursor: pointer;
-        display: flex;
-        align-items: center;
-        margin: 5px;
-    }
-
-    .rail {
-        width: 100%;
-        height: 8px;
-        background-color: #000000;
-        border: 1px solid #888;
-        border-radius: 8px;
-        overflow: hidden;
-    }
-
-    .fill {
-        width: 100%;
-        height: 8px;
-        background-color: #d6d9c7;
-        transform-origin: left;
-        transform: scaleX(0);
-    }
-
-    .ball {
-        position: absolute;
-        z-index: 2;
-        top: calc(50% - 6.5px);
-        left: -3px;
-        width: 13px;
-        height: 13px;
-        border-radius: 50%;
-        background-color: #968482;
-    }
-
     button {
         border: none;
         padding: 0;
@@ -121,11 +38,6 @@
         background-color: transparent;
     }
 </style>
-
-<svelte:body
-    on:mousemove={handleMouseMove}
-    on:mouseup={handleMouseUp}
-    on:mouseleave={handleMouseUp} />
 
 <div class="slidecontainer">
     <button id="muteUnmute" on:click={() => dispatch('muteUnmute')}>
@@ -138,27 +50,16 @@
         </label>
     </button>
 
-    <div
-        tabindex="0"
-        aria-valuemax={max}
-        aria-valuemin={min}
-        aria-valuenow={current}
-        aria-orientation="horizontal"
-        role="slider"
-        bind:this={slider}
-        class="slider"
-        on:mousedown={handleMouseDown}
-        on:keydown={handleKeyDown}>
-        <div class="rail">
-            <div
-                class="ball"
-                style="left: calc({(current / max) * 100}% - 6.5px)" />
-
-            <div class="fill" style="transform: scaleX({current / max})" />
-        </div>
-    </div>
+    <Slider
+        bind:min
+        bind:max
+        bind:step
+        bind:current
+        {size}
+        {sliderWidth}
+        {fillColor} />
 
     <label for="vol-slider" style="text-align:right">
-        {Math.floor(current * 100)}
+        <div>{Math.floor(current * 100)}</div>
     </label>
 </div>
